Refactor code and add API documentation

+++ Gateway/src/product/product.controller.ts
@@ -1,6 +1,6 @@
-import { Controller, Post, Get, Body, UploadedFile, UseInterceptors, Param, Delete, UseGuards, Request, UploadedFiles } from '@nestjs/common';
+import { Controller, Post, Get, Body, UploadedFile, UseInterceptors, Param, Delete, UseGuards, Request, UploadedFiles, UnauthorizedException, Put } from '@nestjs/common';
 import { ProductService } from './product.service';
-import { ApiParam, ApiBody, ApiConsumes, ApiOkResponse, ApiProperty, ApiTags, ApiBearerAuth, ApiCreatedResponse } from '@nestjs/swagger';
+import { ApiParam, ApiBody, ApiConsumes, ApiOkResponse, ApiProperty, ApiTags, ApiBearerAuth, ApiCreatedResponse, ApiOperation } from '@nestjs/swagger';
 import { CategoriesDto } from './dto/category.dto';
 import { FileFieldsInterceptor, FileInterceptor } from '@nestjs/platform-express';
 import { ProductDto } from './dto/product.dto';
@@ -18,6 +18,8 @@ export class ProductController {
 
     @Post('category')
     @Roles(UserRole.ADMIN)
+    @ApiOperation({ summary: 'Create a category (admin required)' })
+    @ApiTags('Category', 'Admin')
     @ApiConsumes('multipart/form-data')
     @UseInterceptors(FileInterceptor('image'))
     @ApiBody({ schema: {
@@ -54,19 +56,61 @@ export class ProductController {
     }
 
     @Get('category')
+    @ApiOperation({ summary: 'Get all categories' })
+    @ApiTags('Category')
     @ApiOkResponse({ type: [CategoriesDto] })
     async getAllCategories(): Promise<any> {
         return this.productService.getAllCategories();
     }
 
     @Get('category/:id')
+    @ApiOperation({ summary: 'Get category by id' })
+    @ApiTags('Category')
     @ApiParam({ name: 'id', type: Number })
     @ApiOkResponse({ type: CategoriesDto })
     async getCategoryById(@Param('id') id: number): Promise<any> {
         return this.productService.getCategoryById(id);
     }
+
+    @Put('category/:id')
+    @Roles(UserRole.ADMIN)
+    @ApiOperation({ summary: 'Edit a category (admin required)' })
+    @ApiTags('Category', 'Admin')
+    @ApiParam({ name: 'id', type: Number })
+    @ApiConsumes('multipart/form-data')
+    @UseInterceptors(FileInterceptor('image'))
+    @ApiBody({ schema: {
+        type: 'object',
+        properties: {
+            name: { type: 'string', nullable: true},
+            description: { type: 'string', nullable: true},
+            parent: { type: 'number', nullable: true},
+            image: {
+                type: 'string',
+                format: 'binary',
+                nullable: true
+            }
+        }
+    }
+    })
+    @ApiOkResponse({ type: CategoriesDto })
+    async editCategory(@Param('id') id: number, @Body() param: {
+        name?: string,
+        description?: string,
+        parent?: number,
+    }, @UploadedFile() image: Express.Multer.File): Promise<any> {
+        return this.productService.editCategory({
+            id: id,
+            name: param.name,
+            description: param.description,
+            parent: param.parent,
+            image: image
+        });
+    }
     
     @Post('store')
+    @ApiOperation({ summary: 'Register a store (Become a store owner)' })
+    @ApiTags('Store')
     @UseGuards(JwtAuthGuard)
     @ApiBearerAuth()
     @ApiConsumes('multipart/form-data')
@@ -106,20 +150,31 @@ export class ProductController {
             name: param.name,
             description: param.description,
             address: param.address,
-            logo: files.logo[0] || null,
-            cover: files.cover[0] || null
+            logo: files.logo ? files.logo[0] : null,
+            cover: files.cover ? files.cover[0] : null
         });
     }
 
     @Get('unapproved-store')
     @Roles(UserRole.ADMIN)
+    @ApiOperation({ summary: 'Get all unapproved stores (admin required)' })
+    @ApiTags('Store', 'Admin')
     @ApiOkResponse({ type: [CategoriesDto] })
     async getAllUnapprovedStores(): Promise<StoreDto[]> {
-        console.log('get all unapproved stores')
         return this.productService.getAllUnapprovedStores();
     }
 
+    @Get('store')
+    @ApiOperation({ summary: 'Get all stores' })
+    @ApiTags('Store')
+    @ApiOkResponse({ type: [CategoriesDto] })
+    async getAllStores(): Promise<StoreDto[]> {
+        return this.productService.getAllApprovedStores();
+    }
+
     @Get('store/:id')
+    @ApiOperation({ summary: 'Get store by id' })
+    @ApiTags('Store')
     @ApiParam({ name: 'id', type: Number })
     @ApiOkResponse({ type: CategoriesDto })
     async getStoreById(@Param('id') id: number): Promise<StoreDto> {
@@ -129,6 +184,8 @@ export class ProductController {
 
     @Post('store/approve/:id')
     @Roles(UserRole.ADMIN)
+    @ApiOperation({ summary: 'Approve a store (admin required)' })
+    @ApiTags('Store', 'Admin')
     @ApiParam({ name: 'id', type: Number })
     @ApiOkResponse({ type: CategoriesDto })
     async approveStore(@Param('id') id: number): Promise<StoreDto> {
@@ -139,6 +196,8 @@ export class ProductController {
     @Post()
     @ApiBearerAuth()
     @UseGuards(JwtAuthGuard)
+    @ApiOperation({ summary: 'Create a product (store owner required)' })
+    @ApiTags('Product', 'Store owner')
     @ApiConsumes('multipart/form-data')
     @UseInterceptors(FileInterceptor('image'))
     @ApiBody({ schema: {
@@ -171,22 +230,80 @@ export class ProductController {
     }
 
     @Get(':id')
+    @ApiOperation({ summary: 'Get product by id' })
+    @ApiTags('Product')
     @ApiParam({ name: 'id', type: Number })
     @ApiOkResponse({ type: ProductDto })
     async getProductById(@Param('id') id: number): Promise<any> {
         return this.productService.getProductById(id);
     }
 
-    @Delete(':id')
-    @Roles(UserRole.ADMIN)
+    @Put(':id')
+    @ApiOperation({ summary: 'Edit a product (store owner required)' })
+    @ApiTags('Product', 'Store owner')
+    @ApiBearerAuth()
+    @UseGuards(JwtAuthGuard)
+    @ApiConsumes('multipart/form-data')
+    @UseInterceptors(FileInterceptor('image'))
+    @ApiBody({ schema: {
+        type: 'object',
+        properties: {
+            name: { type: 'string', nullable: true},
+            description: { type: 'string', nullable: true},
+            category: { type: 'number', nullable: true},
+            image: {
+                type: 'string',
+                format: 'binary',
+                nullable: true
+            }
+        }
+    }})
     @ApiParam({ name: 'id', type: Number })
     @ApiOkResponse({ type: ProductDto })
-    async removeProductById(@Param('id') id: number): Promise<any> {
+    async editProduct(@Request() req, @Param('id') id: number, @Body() param: {
+        name?: string,
+        description?: string,
+        category?: number,
+    }, @UploadedFile() image: Express.Multer.File): Promise<any> {
+        const storeId = req.user.id;
+        if (!(await this.productService.storeHas({
+            storeId: storeId,
+            productId: id
+        }))) {
+            throw new UnauthorizedException('This product is not in your store');
+        }
+        return this.productService.editProduct({
+            id: id,
+            name: param.name,
+            description: param.description,
+            category: param.category,
+            image: image
+        });
+    }
+
+    @Delete(':id')
+    @ApiOperation({ summary: 'Remove a product (store owner required)' })
+    @ApiTags('Product', 'Store owner')
+    @ApiBearerAuth()
+    @UseGuards(JwtAuthGuard)
+    @ApiParam({ name: 'id', type: Number })
+    @ApiOkResponse({ type: ProductDto })
+    async removeProductById(@Request() req, @Param('id') id: number): Promise<any> {
+        const storeId = req.user.id;
+        if (!(await this.productService.storeHas({
+            storeId: storeId,
+            productId: id
+        }))) {
+            throw new UnauthorizedException('This product is not in your store');
+        }
         return this.productService.removeProductById(id);
     }
 
     @Post('variant/:product/:size/:color')
-    @Roles(UserRole.ADMIN)
+    @ApiOperation({ summary: 'Set a product variant (store owner required)' })
+    @ApiTags('Product', 'Store owner')
+    @ApiBearerAuth()
+    @UseGuards(JwtAuthGuard)
     @ApiConsumes('multipart/form-data')
     @UseInterceptors(FileInterceptor('image'))
     @ApiBody({ schema: {
@@ -204,10 +321,17 @@ export class ProductController {
     @ApiParam({ name: 'product', type: Number })
     @ApiParam({ name: 'size', type: String })
     @ApiParam({ name: 'color', type: String })
-    async setProductVariant(@Param('product') product: number, @Param('size') size: string, @Param('color') color: string, @Body() param: {
+    async setProductVariant(@Request() req, @Param('product') product: number, @Param('size') size: string, @Param('color') color: string, @Body() param: {
         price?: number,
         quantity?: number,
     }, @UploadedFile() image: Express.Multer.File): Promise<any> {
+        const storeId = req.user.id;
+        if (!(await this.productService.storeHas({
+            storeId: storeId,
+            productId: product
+        }))) {
+            throw new UnauthorizedException('This product is not in your store');
+        }
         return this.productService.setProductVariant({
             product: product,
             size: size,
@@ -219,6 +343,8 @@ export class ProductController {
     }
 
     @Get('variant/:productId/:size/:color')
+    @ApiOperation({ summary: 'Get a product variant' })
+    @ApiTags('Product')
     @ApiParam({ name: 'productId', type: Number })
     @ApiParam({ name: 'size', type: String })
     @ApiParam({ name: 'color', type: String })
@@ -232,12 +358,22 @@ export class ProductController {
     }
 
     @Delete('variant/:productId/:size/:color')
-    @Roles(UserRole.ADMIN)
+    @ApiOperation({ summary: 'Remove a product variant (store owner required)' })
+    @ApiTags('Product', 'Store owner')
+    @ApiBearerAuth()
+    @UseGuards(JwtAuthGuard)
     @ApiParam({ name: 'productId', type: Number })
     @ApiParam({ name: 'size', type: String })
     @ApiParam({ name: 'color', type: String })
     // @ApiOkResponse({ type: ProductDto })
-    async removeProductVariant(@Param('productId') productId: number, @Param('size') size: string, @Param('color') color: string): Promise<any> {
+    async removeProductVariant(@Request() req, @Param('productId') productId: number, @Param('size') size: string, @Param('color') color: string): Promise<any> {
+        const storeId = req.user.id;
+        if (!(await this.productService.storeHas({
+            storeId: storeId,
+            productId: productId
+        }))) {
+            throw new UnauthorizedException('This product is not in your store');
+        }
         return this.productService.removeProductVariant({
             productId: productId,
             size: size,
